Moved linux methods into CommandKit::OS::Linux.
